Implemented Date Labeling Functionality & Debugging
Added dates to each entry in the output. This required extensive retooling of the displaySchedulePreview() and the creation of the prepareSchedulePreview(). A third array list also had to be implemented to hold the number of days to read for each book in order for the prepareSchedulePreview() to not distort the number of days for the prior book. Also squashed several bugs in preparation for the testing phase. This preliminarily concludes the main development phase of the program.